--- a/spec/unit/fetchers/security_group_list_fetcher_spec.rb
+++ b/spec/unit/fetchers/security_group_list_fetcher_spec.rb
@@ -11,16 +11,40 @@ module VCAP::CloudController
     let!(:security_group_1) { SecurityGroup.make }
     let!(:security_group_2) { SecurityGroup.make }
     let!(:security_group_3) { SecurityGroup.make }
+    let(:security_groups) {}
+    let(:associated_space) { Space.make }
+
+    shared_examples 'eager loading' do
+      it 'eager loads running and staging spaces' do
+        security_groups.all.each { |sg| expect(sg.associations.keys).to contain_exactly(:spaces, :staging_spaces) }
+      end
+
+      it 'eager loads space guids only' do
+        security_group_1.add_space(associated_space)
+        security_group_1.add_staging_space(associated_space)
+        associations = security_groups.where(guid: security_group_1.guid).all.first.associations
+        [:spaces, :staging_spaces].each do |key|
+          expect(associations[key].length).to eq(1)
+          expect(associations[key].first.keys).to contain_exactly(:guid)
+        end
+      end
+    end
 
     context '#fetch_all' do
+      let(:security_groups) { fetcher.fetch_all(message) }
+
+      include_examples 'eager loading'
+
       it 'includes all the security_groups' do
-        expect(fetcher.fetch_all(message).all).to include(security_group_1, security_group_2, security_group_3)
+        expect(security_groups.all).to include(security_group_1, security_group_2, security_group_3)
       end
     end
 
     describe '#fetch' do
       let(:visible_security_groups) { [security_group_1.guid, security_group_2.guid] }
       let(:security_groups) { fetcher.fetch(message, visible_security_groups) }
+
+      include_examples 'eager loading'
 
       context 'when no filters are specified' do
         it 'returns all of the security groups' do
@@ -43,7 +67,6 @@ module VCAP::CloudController
       end
 
       context 'when we filter on running_space_guid' do
-        let(:associated_space) { Space.make }
         let(:filters) { { running_space_guids: [associated_space.guid] } }
 
         before do
@@ -56,7 +79,6 @@ module VCAP::CloudController
       end
 
       context 'when we filter on staging_space_guid' do
-        let(:associated_space) { Space.make }
         let(:filters) { { staging_space_guids: [associated_space.guid] } }
 
         before do