This PR is a follow up for #2255 and further improves the performance of /v3/security_groups endpoints. Running and staging space guids are eagerly loaded (only the guids by modifying the dataset used for eager loading). The same is done when a single security group is fetched. Furthermore permission checks are done as part of the where clauses.

Here are some results of locally executed performance tests:

|User|Request|X-Runtime before change|X-Runtime after change|
|--|--|--|--|
|as admin|GET /v3/security_groups?per_page=500 \*|5.520s|3.492s|
|as regular user|GET /v3/security_groups/:guid|0.158s|0.029s|

\* each security group is assigned to 500 spaces

* [x] I have reviewed the [contributing guide](https://github.com/cloudfoundry/cloud_controller_ng/blob/main/CONTRIBUTING.md)

* [x] I have viewed, signed, and submitted the Contributor License Agreement

* [x] I have made this pull request to the `main` branch

* [x] I have run all the unit tests using `bundle exec rake`

* [ ] I have run [CF Acceptance Tests](https://github.com/cloudfoundry/cloud_controller_ng#cf-acceptance-tests-cats)