@JenGoldstrich 2021-09-15
Deployed these changes to a bosh lite with ~100 security groups and validated that curling /v3/security_groups returned in less than a second, none of them bound mind you. 

Looks good to merge but @philippthun can you resolve the merge conflict above and then I will happily merge this!